added VERSION file
The VERSION file is changed automatically when git-archive is called (i.e.
downloading zip from website). This is specified in the .gitattributes.

If the repository is cloned then the VERSION file is changed by
the update_version.sh script which is called by make install (GNUmakefile).

The VERSION file should be changed only when bumping the imunes VERSION and just
the first line.

Bump VERSION to 2.0 because of introduced Linux support.