added root check and disabled devicemapper in hcp
Copying from and betweeen devicemapper virtual nodes doesn't work.
It only works when copying files from nodes.

We're not currently considering implementing this. If we were to do this we
should mount devicemapper devices in order to copy files to the drives.
Docker runs slow on devicemapper and therefore we don't see any point in
supporting this option in hcp for Linux.