added new vlink.tcl script
For now works just on Linux. For it to work on FreeBSD, 2 procedures must be
populated (getFreeBSDLinkStatus and applyLinkSettingsFreeBSD).

This script relies on IMUNES to create a file with needed link data in
/var/run/imunes/_eid_/links. This is needed to avoid magically detecting which
links are avaliable, and greatly reduces the amount of logic needed to check
whether a link exists or not. It also makes the detection process OS
independent.
Added two new features: print all avaliable links and print link status.