preparation for packaging (dynamic ROOTDIR)
Using PREFIX when running make install now works, for example:
 # make PREFIX=/usr install

When using PKGBUILD on Arch Linux, the installation is done in a fakeroot
environment, so the PREFIX from Makefile (/usr/local) is overwritten. This new
PREFIX should then be forwarded to update_version.sh, otherwise the script
would use the hardcoded ROOTDIR variable (/usr/local) and the installation
would fail.