added FreeBSD pkg package files
This is experimental and not tested enough (but it worked on couple of my
computers so I'm uploading it as a starting point).

These files are used to create the imunes package on a FreeBSD system. Run:
 # make package
to create package in work/pkg/ folder.
To install the package, run:
 # pkg install imunes*.txz

The port downloades the master.tar.gz from imunes github master branch,
extracts it to work folder, builds it in work/stage/ and uses some additional
scripts to install/deinstall /var/imunes/vroot/ virtual filesystem.