Revive a long forgotten wlan node
FreeBSD only.

How to test it? Suppose we want to connect an existing router node and
an existing pc node.
1. Add a wlan node.
2. Select a link tool.
3. Make a link between pc node and a wlan node (NOTE: the link won't be
visible. Nevertheless, the wlan interface on the pc node is created and
can be seen in the statusline at the bottom of GUI when the mouse hovers
above the pc node.)
4. Make a link between router node and a wlan node (like in step 3.)
5. Run experiment.
6. The 'strength' of the link depends on the closeness between the pc
and router nodes.

There are a lot of things that can be improved here, starting with
customizing wifi link parameters in GUI (dup, BER, BW), showing the link
strength somewhere in GUI...