Reload topogen.tcl and debug.tcl on theme change
Changing the theme proved to be a much larger quest than I thought it
would be. The menubar background/foreground colors cannot be changed by
configuring the it live - it needs to be destroyed and redrawn again.
However, this introduces additional complications: all the button states
are reset to default, requiring us to retain the states and dynamically
adjust the buttons... Of course, this can be done, but we need a bigger
GUI update than I thought. I'll get back to this in the future, maybe.

This is all unnecessary with the default (light) themes, as they are all
alike and share colors.