Fix static routes when linking to routers/extnats
Up until now, when the user connected a node to a router node, IMUNES
would assign only one default route (the router which was connected
last). Now we append a default route for each router (and extnat). That
way, when we delete a link to any of them, we could also remove the
route (if it exists).