Separate user static routes and auto default ones
Automatic default routes are created for each interface dynamically on
experiment startup and appended to node startup configuration. It is
possible to have multiple default routes for each interface of the node.
This means that they can be removed when the interface is removed and
all the other default routes (and user static routes) are untouched.

They are visible in GUI configuration along with a checkbutton to
enable/disable automatically assigning routes on experiment startup.

PC and host nodes get automatic default routes for each router and
extnat nodes in the same L2 network. Router nodes get the automatic
default route for each extnat in the same L2 network.