Dual support for svg/gif
I would like to keep gif support for legacy reasons. The svg
functionality is included in tk8.7, but we need tksvg if tk8.6 is used.
If tksvg is not installed, we use gifs for GUI icons.

When (if?) tk8.7 is released, I think we'll keep supporting 8.6 for
quite some time.

TODO: Not really tested yet.