Add a 'direct link' config option (no ng_pipe)
Experimental, FreeBSD only! This option is ignored in Linux.

This cannot be enabled in GUI (yet) but can be set as a link option in
.imn files, for example:
link l0 {
 direct 1
 ...
}

When the link is created as a direct_link, all other configurable link
options are ignored.

The default is 0, i.e. ng_pipe is used and link can be configured as
before.