Change the way links are created in Linux
Before creating links, every node creates its own 'physical' interfaces
that will be used to attach to newly created links and do not need to
disappear when links are deleted. This is done using veth pairs - one
half (private hook) is attached to a node netns, and the other (public
hook) is attached to an experiment netns.

Every 'normal' link is a Linux bridge inside an experiment netns with
the name of the link. The public hooks are then connected to their
matching links.

When 'direct' links are introduced, creation of links will be skipped
and veth hooks will be directly connected to their nodes.