Add a 'direct link' option (no ng_pipe/bridge)
This can be enabled in GUI (by right clicking on links) and set as a
link option in .imn files, for example:
link l0 {
 direct 1
 ...
}

When the link is created as a direct_link, all other configurable link
options are ignored.

The default is 0, i.e. ng_pipe/bridge is used and link can be configured
as before.