Remove unneeded ngnodemap array in FreeBSD
We now use a shell variable to save the name of the newly created ng
interface and rename it immediately to something unique inside our
pipesExec call. This means we don't need to wait for the shell to return
in order to know the name of the interface, so we can parallelize this
step.

At the same time, we moved this step away from node initialization, so
we can have better control over it. Currently, this means nothing on
Linux, but there will also be some changes there, as we will want
docker-compose container creation to be independent from the creation of
interfaces.

Moreover, we currently create interfaces on Linux only when creating
links (as veth pairs), so this will change things a bit, and will help
when we introduce creating interfaces without links.