Fix starting FRR for FreeBSD
Replace frrboot.sh with the (slightly modified) quaggaboot.sh as we
otherwise need to wait for a long time to start all the daemons.

We do this just like we did with quagga: each daemon separately,
depending on which of them are in use in our configuration.

Maybe there's a better way, but stick with this for now.