Enable editing topologies in exec mode
WARNING: not a stable commit

This enables changing the topology while the experiment is running.
Additionally, it is possible to pause/resume execution of new
nodes/ifaces/links using the Pause/Resume experiment options.

-> Introduce trigger_* procedures

Used to trigger nodes/links/interfaces creation, destruction,
recreation, or configuration, unconfiguration or reconfiguration.

When custom config is enabled or when auto default routes is disabled on
node, do not automatically reconfigure them. This would usually happen
when connecting/disconnecting a router/extnat node in the same subnet or
when changing some router/extnats interface IP address.

Also, do not reconfigure if custom config is selected and the user
toggles auto default routes in node config.