Separate interfaces and node custom configurations
WARNING: not a stable commit

Instead of having only a single custom configuration per node, it is now
possible to use hooks for different events in node creation. Currently,
those are IFACES_CONFIG and NODE_CONFIG, so we can have separate custom
configurations for interfaces and for the node. This means we can
disable interfaces custom configuration and let the IMUNES configure the
interfaces with default values and modify node configuration script
independently.