Add a 3s timeout on node/ifaces creation/config
If a node is not created (jail or docker container), we skip other steps
in its execution.
Similarly, if a node/ifaces configurations take too long, we skip
checking for errors.

Currently, timeout is not modifiable from GUI (only from imunes.tcl) -
there will be an option to change it in the future.