Upgrade to a new JSON .imn format
Convert old .imn to JSON when opening old formats (use loadCfgLegacy
proc for that). New JSON format will be loaded using a new procedure
(loadCfgJson). It is only possible to save the new format.

Additionally, config versioning is enabled, in case of configuration
changes and to handle compatibility a bit easier.