Merge rj45 and extelem nodes
WARNING: not a stable commit

This means that the existing rj45 nodes will be kept as they were, but
extelem nodes will be shown as rj45 nodes with multiple interfaces in
tabbed GUI config.

This also enables multiple stolen interfaces of the same node to be vlan
tagged (previously not possible with extelem).