Enable control over (un)supported node types
It is now possible to mark node types 'runnable' in order for them to be
shown in the left toolbar menu. A new View menu is added which toggles
whether unsupported node types are shown in the left toolbar menu.

The redrawing of left toolbar has been refactored into a procedure in
order to be able to call it if needed.

It was also moved to drawing.tcl in order to be available when
resourcing files in debug mode.