Extend IMUNES rc config files
Read the first configuration file that exists, in this order:
 ./.imunesrc
 ./.imunes.rc
 $HOME/.imunes.rc
 $XDG_CONFIG_HOME/imunes/config
 /etc/imunes/config

After that, read /etc/imunes/override - options set here will override
any previously set options.

All the configuration files are JSON files, except for .imunesrc, which
is a TCL file which will be sourced. It is kept for compatibility
reasons.

Check config.example for the list of options.