Add new parameters for 'legacy mode'
Run with:
$ imunes -l
or
$ imunes -legacy

Legacy mode changes just the way we check if nodes are
started/configured (as explained in the last commit).

This is useful only for slow computers/VMs running Linux, or for VERY
big topologies.

NOTE:
Because node creation/configuration will not time out, there is a
possibility that the experiment will get stuck trying to run/configure a
single node until the user kills the experiment, leaving the other nodes
still running and leaving experiment in a non-finished state. This can
happen if a node is misconfigured, so a user needs to clean the
namespaces, jails/dockers and experiment files manually.
Running in non-legacy mode stops this from happening, so avoid legacy
mode unless nodes are properly configured.