Refactor printing to file
And add 'Print to file' hack for better output.

Annotation texts need to be smaller when printing the canvas for some
reason (italic text use a different factor). That's why we lower it
before printing, then get them back after printing them.

Also, some of my fonts are saved with the wrong name to the .ps file
(missing the '-Regular' suffix), so fix this before converting to .pdf.